+++ core/src/main/java/edu/umich/verdict/query/CreateTableQuery.java
@@ -8,8 +8,8 @@ import java.util.Map;
 import org.apache.commons.lang3.tuple.Pair;
 
 import edu.umich.verdict.VerdictJDBCContext;
-import edu.umich.verdict.VerdictSQLBaseVisitor;
-import edu.umich.verdict.VerdictSQLParser;
+import edu.umich.verdict.parser.VerdictSQLBaseVisitor;
+import edu.umich.verdict.parser.VerdictSQLParser;
 import edu.umich.verdict.exceptions.VerdictException;
 import edu.umich.verdict.util.VerdictLogger;
