+++ core/src/main/java/edu/umich/verdict/query/DescribeTableQuery.java
@@ -1,9 +1,9 @@
 package edu.umich.verdict.query;
 
 import edu.umich.verdict.VerdictContext;
-import edu.umich.verdict.VerdictSQLBaseVisitor;
-import edu.umich.verdict.VerdictSQLParser;
-import edu.umich.verdict.VerdictSQLParser.Table_nameContext;
+import edu.umich.verdict.parser.VerdictSQLBaseVisitor;
+import edu.umich.verdict.parser.VerdictSQLParser;
+import edu.umich.verdict.parser.VerdictSQLParser.Table_nameContext;
 import edu.umich.verdict.datatypes.TableUniqueName;
 import edu.umich.verdict.dbms.DbmsJDBC;
 import edu.umich.verdict.dbms.DbmsSpark;
