+++ core/src/main/java/edu/umich/verdict/relation/ExactRelation.java
@@ -11,13 +11,13 @@ import org.apache.commons.lang3.tuple.Triple;
 import com.google.common.base.Optional;
 
 import edu.umich.verdict.VerdictContext;
-import edu.umich.verdict.VerdictSQLBaseVisitor;
-import edu.umich.verdict.VerdictSQLParser;
-import edu.umich.verdict.VerdictSQLParser.Group_by_itemContext;
-import edu.umich.verdict.VerdictSQLParser.Join_partContext;
-import edu.umich.verdict.VerdictSQLParser.Order_by_expressionContext;
-import edu.umich.verdict.VerdictSQLParser.Select_list_elemContext;
-import edu.umich.verdict.VerdictSQLParser.Table_sourceContext;
+import edu.umich.verdict.parser.VerdictSQLBaseVisitor;
+import edu.umich.verdict.parser.VerdictSQLParser;
+import edu.umich.verdict.parser.VerdictSQLParser.Group_by_itemContext;
+import edu.umich.verdict.parser.VerdictSQLParser.Join_partContext;
+import edu.umich.verdict.parser.VerdictSQLParser.Order_by_expressionContext;
+import edu.umich.verdict.parser.VerdictSQLParser.Select_list_elemContext;
+import edu.umich.verdict.parser.VerdictSQLParser.Table_sourceContext;
 import edu.umich.verdict.datatypes.SampleParam;
 import edu.umich.verdict.datatypes.TableUniqueName;
 import edu.umich.verdict.exceptions.VerdictException;
