+++ core/src/main/java/edu/umich/verdict/relation/expr/Expr.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import edu.umich.verdict.VerdictContext;
-import edu.umich.verdict.VerdictSQLBaseVisitor;
-import edu.umich.verdict.VerdictSQLParser;
+import edu.umich.verdict.parser.VerdictSQLBaseVisitor;
+import edu.umich.verdict.parser.VerdictSQLParser;
 import edu.umich.verdict.util.StringManipulations;
 
 public abstract class Expr {
