+++ core/src/main/java/edu/umich/verdict/relation/expr/OrderByExpr.java
@@ -2,8 +2,8 @@ package edu.umich.verdict.relation.expr;
 
 import com.google.common.base.Optional;
 
-import edu.umich.verdict.VerdictSQLBaseVisitor;
-import edu.umich.verdict.VerdictSQLParser;
+import edu.umich.verdict.parser.VerdictSQLBaseVisitor;
+import edu.umich.verdict.parser.VerdictSQLParser;
 import edu.umich.verdict.util.StringManipulations;
 
 public class OrderByExpr extends Expr {
