+++ pom.xml
@@ -112,7 +112,6 @@
 								<goal>aggregate</goal>
 							</goals>
 							<phase>site</phase>
-
 							<configuration>
 								<!-- Specific configuration for the aggregate report -->
 							</configuration>
